+++ app/src/main/java/de/k3b/android/androFotoFinder/Global.java
@@ -36,6 +36,8 @@ public class Global {
     public static boolean debugEnabledViewItem = false;
     public static boolean debugEnabledSql = true;
 
+    public static int maxSelectionMarkersInMap = 255;
+
     public static boolean clearSelectionAfterCommand = false;
 
     public static void debugMemory(String modul, String message) {